These size_t arguments should be, well, according to POSIX, socklen_t, but Be declared them as int. So do I.
git-svn-id: file:///srv/svn/repos/haiku/trunk/current@8567 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ headers/posix/sys/socket.h
@@ -273,8 +273,8 @@ int     shutdown(int sock, int how);
 
 ssize_t send(int, const void *, size_t, int);
 ssize_t recv(int, void *, size_t, int);
-ssize_t sendto(int, const void *, size_t, int, const struct sockaddr *, size_t);
-ssize_t recvfrom(int, void *, size_t, int, struct sockaddr *, size_t *);
+ssize_t sendto(int, const void *, size_t, int, const struct sockaddr *, int);
+ssize_t recvfrom(int, void *, size_t, int, struct sockaddr *, int *);
 
 int     setsockopt(int, int, int, const void *, size_t);
 int     getsockopt(int, int, int, void *, size_t *);
